pyproject.toml
[project]
name = "mcp-adapters"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.3",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
]
